Fix initial layers settings state
